Rename NotifyService folder
