+++ src/SFA.DAS.Notifications.Infrastructure/GovNotifyService/GovNotifyPayload.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace SFA.DAS.Notifications.Infrastructure.NotifyEmailService
+namespace SFA.DAS.Notifications.Infrastructure.GovNotifyService
 {
     public class GovNotifyPayload
     {
