+++ src/SFA.DAS.Notifications.Infrastructure/GovNotifyService/NotifyMessage.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
 
-namespace SFA.DAS.Notifications.Infrastructure.NotifyEmailService
+namespace SFA.DAS.Notifications.Infrastructure.GovNotifyService
 {
     public class NotifyMessage
     {
