+++ src/SFA.DAS.Notifications.Infrastructure/GovNotifyService/NotifySmsService.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using SFA.DAS.Notifications.Application.Interfaces;
 using SFA.DAS.Notifications.Domain.Entities;
 
-namespace SFA.DAS.Notifications.Infrastructure.NotifyEmailService
+namespace SFA.DAS.Notifications.Infrastructure.GovNotifyService
 {
     public class NotifySmsService : ISmsService
     {
